휴가신청 & 취소 & 상태조회 개발 중

+++ src/main/java/com/example/demo/entity/MemberVacationM.java
@@ -5,6 +5,7 @@ import lombok.Getter;
 import lombok.Setter;
 
 import javax.persistence.*;
+import java.math.BigDecimal;
 
 @Entity
 @Getter @Setter
@@ -23,8 +24,8 @@ public class MemberVacationM {
     private String vacationYear;
 
     @Column
-    private Double totalCount;
+    private BigDecimal totalCount;
 
     @Column
-    private Double useCount;
+    private BigDecimal useCount;
 }
